Work on fixing tablet view of results, keeping current PanelView component
