+++ packages/esm-patient-tests-app/src/index.ts
@@ -53,10 +53,7 @@ export const addLabOrderWorkspace = getAsyncLifecycle(
   options,
 );
 
-export const timelineResultsModal = getAsyncLifecycle(
-  () => import('./test-results/panel-timeline/timeline-results.modal'),
-  {
-    featureName: 'Timeline results',
-    moduleName,
-  },
-);
+export const timelineResultsModal = getAsyncLifecycle(() => import('./test-results/trendline/timeline-results.modal'), {
+  featureName: 'Timeline results',
+  moduleName,
+});
