+++ packages/esm-patient-tests-app/src/test-results/helpers.ts
@@ -1,4 +1,3 @@
-import { navigate } from '@openmrs/esm-framework';
 import { dashboardMeta } from './dashboard.meta';
 
 export const makeThrottled = <T extends (...args: any[]) => any>(
@@ -28,15 +27,5 @@ export const makeThrottled = <T extends (...args: any[]) => any>(
 
   return throttledFunc;
 };
-
-export const navigateToTimeline = (basePath: string, panelUuid: string) =>
-  navigate({ to: `${testResultsBasePath(basePath)}/timeline/${panelUuid}` });
-
-export const navigateToTrendline = (basePath: string, panelUuid: string, testUuid: string) =>
-  navigate({
-    to: `${testResultsBasePath(basePath)}/trendline/${panelUuid}/${testUuid}`,
-  });
-
-export const navigateToResults = (basePath: string) => navigate({ to: `${testResultsBasePath(basePath)}` });
 
 export const testResultsBasePath = (basePath: string) => `${window.spaBase}${basePath}/${dashboardMeta.path}`;
