--- packages/esm-patient-tests-app/src/test-results/panel-timeline/timeline.scss
@@ -1,224 +1,0 @@
-@use '@carbon/colors';
-@use '@carbon/layout';
-@use '@carbon/type';
-@use '@openmrs/esm-styleguide/src/vars' as *;
-
-.grid {
-  width: fit-content;
-  display: grid;
-  grid-auto-rows: auto;
-  gap: 0;
-  justify-items: center;
-  align-items: center;
-
-  > div {
-    border: 0.5px solid colors.$gray-50;
-    border-bottom: none;
-    border-left: none;
-  }
-}
-
-.grid:hover {
-  border: none;
-}
-
-.day-column,
-.year-column {
-  @include type.type-style('heading-compact-01');
-  color: $text-02;
-}
-
-.time-column {
-  @include type.type-style('body-compact-01');
-  color: $text-02;
-  scroll-snap-align: start;
-}
-
-.padded-main {
-  box-sizing: border-box;
-  padding: 20px;
-}
-
-.timeline-cell {
-  background-color: white;
-  width: 100%;
-  height: 100%;
-  display: grid;
-  justify-items: center;
-  align-items: center;
-  padding: layout.$spacing-03;
-}
-
-.timeline-data-cell {
-  @extend .timeline-cell;
-  justify-items: first baseline;
-
-  p {
-    @include type.type-style('body-compact-01');
-    color: $text-02;
-  }
-
-  &:hover {
-    background-color: colors.$blue-10;
-    border-top: 2px colors.$blue-30;
-  }
-}
-
-.row-start-cell {
-  @extend .timeline-cell;
-  position: sticky;
-  left: 0;
-  display: grid;
-  grid-auto-flow: row;
-  justify-items: baseline;
-  align-items: center;
-  padding: layout.$spacing-05;
-}
-
-.timeline-cell-zebra {
-  background-color: $ui-03;
-}
-
-.padding-container {
-  padding: 0px 0 0 0;
-  box-sizing: border-box;
-  display: grid;
-  grid-auto-flow: row;
-  row-gap: 0px;
-  grid-template: auto auto / 9em auto;
-  border-collapse: collapse;
-  width: fit-content;
-  max-width: 100%;
-}
-
-.padding-container::-webkit-scrollbar {
-  display: none;
-}
-
-.time-slot-inner {
-  background-color: $ui-03;
-  padding: 1px 10px;
-  justify-self: stretch;
-  align-self: stretch;
-  display: grid;
-  align-items: center;
-
-  span {
-    position: sticky;
-    left: calc(10em + 10px);
-    width: max-content;
-  }
-}
-
-.corner-grid-element {
-  grid-row: span 1;
-  position: sticky;
-  left: 0px;
-  top: 0px;
-  z-index: 3;
-  border-right: 0.5px solid colors.$gray-30;
-
-  span {
-    @include type.type-style('heading-compact-01');
-    width: auto;
-    left: 0;
-  }
-}
-
-.shadow {
-  box-shadow: layout.$spacing-03 0 20px 0 rgba(0, 0, 0, 0.15);
-}
-
-.trendline-link {
-  @include type.type-style('body-compact-01');
-  cursor: pointer;
-}
-
-.range-units {
-  @include type.type-style('helper-text-01');
-  color: colors.$gray-70;
-}
-
-.off-scale-high,
-.off-scale-low,
-.critically-high,
-.critically-low,
-.high,
-.low {
-  p {
-    @include type.type-style('heading-compact-01');
-  }
-}
-
-.high,
-.low {
-  background-color: colors.$orange-10;
-  border-top: 1px solid colors.$orange-20;
-}
-
-.critically-high,
-.critically-low {
-  background-color: colors.$red-20;
-  border-top: 1px solid colors.$red-20-hover;
-}
-
-.off-scale-low {
-  p::after {
-    content: ' ↓↓↓';
-  }
-}
-
-.off-scale-high {
-  p::after {
-    content: ' ↑↑↑';
-  }
-}
-
-.critically-low {
-  p::after {
-    content: ' ↓↓';
-  }
-}
-
-.critically-high {
-  p::after {
-    content: ' ↑↑';
-  }
-}
-
-.low {
-  p::after {
-    content: ' ↓';
-  }
-}
-
-.high {
-  p::after {
-    content: ' ↑';
-  }
-}
-
-.shadow-box {
-  grid-row: 2 / -1;
-  grid-column: 1 / 2;
-}
-
-.recent-results-grid {
-  display: grid;
-  overflow: auto;
-  max-height: calc(100vh - 9rem);
-}
-
-.recent-results-grid::-webkit-scrollbar {
-  display: none;
-}
-
-.trendline-link-view {
-  @include type.type-style('helper-text-01');
-  color: colors.$blue-60;
-  cursor: pointer;
-}
-
-.trendline-link-view:hover {
-  text-decoration: underline;
-}
