--- packages/esm-patient-tests-app/src/test-results/panel-timeline/types.ts
@@ -1,8 +1,0 @@
-import { type ParsedTimeType } from '../filter/filter-types';
-import { type ObsRecord } from '../../types';
-
-export interface TimelineData {
-  parsedTimes: ParsedTimeType;
-  timelineData: Record<string, Array<ObsRecord>>;
-  panelName: string;
-}
