+++ packages/esm-patient-tests-app/src/test-results/results-viewer/results-viewer.scss
@@ -3,6 +3,10 @@
 @use '@carbon/type';
 @use '@openmrs/esm-styleguide/src/vars' as *;
 
+.overlay {
+  padding: layout.$spacing-05;
+}
+
 .resultsContainer {
   padding: 0;
   width: 100%;
