+++ packages/esm-patient-tests-app/translations/en.json
@@ -14,7 +14,6 @@
   "closeSearchBar": "Close search",
   "data": "data",
   "dataLoadError": "Data load error",
-  "dataTimeline": "Data Timeline",
   "dataTimelineText": "Data timeline",
   "date": "Date",
   "dateCollected": "Displaying date collected",
@@ -45,7 +44,6 @@
   "orderReason": "Order reason",
   "orderReasonRequired": "Order reason is required",
   "overTime": "Over time",
-  "panel": "Panel",
   "panels": "panels",
   "pleaseRequiredFields": "Please fill all required fields",
   "print": "Print",
@@ -90,10 +88,7 @@
   "tests": "Tests",
   "testType": "Test type",
   "testTypeRequired": "Test type is required",
-  "timeline": "Timeline",
-  "timelineData_lower": "timeline data",
   "tree": "Tree",
-  "trend": "Trend",
   "trendline": "Trendline",
   "trendlineRangeSelector1Day": "1 day",
   "trendlineRangeSelector1Month": "1 month",
